+++ src/index.jsx
@@ -2,7 +2,7 @@ import { BrowserRouter as Router, Route, Switch } from 'react-router-dom'
 import React from 'react'
 import { createRoot } from 'react-dom/client'
 
-import { LoginCallback, Artists } from 'views'
+import { LoginCallback, SearchArtists } from 'views'
 import { PrivateRoute } from 'components'
 import App from 'App'
 import registerServiceWorker from 'registerServiceWorker'
@@ -13,7 +13,7 @@ root.render(
   <Router>
     <Switch>
       <Route path="/login/callback" component={LoginCallback} />
-      <PrivateRoute path="/artistas" component={Artists} />
+      <PrivateRoute path="/artistas" component={SearchArtists} />
       <PrivateRoute path="/" component={App} />
     </Switch>
   </Router>,
