+++ src/views/index.js
@@ -1,3 +1,3 @@
 export { default as Home } from './Home'
 export { default as LoginCallback } from './LoginCallback'
-export { default as Artists } from './Artists'
+export { default as SearchArtists } from './SearchArtists'
